refactor: move callACtionWithIndicesAsync to cad-model
