+++ viewer/packages/cad-model/src/callActionWithIndicesAsync.ts
